Reviewing code and comments

+++ src/main/java/util/Time.java
@@ -1,12 +1,12 @@
 package util;
 
 public class Time {
-	public static float timeStarted = System.nanoTime();
+	public static long timeStarted = System.nanoTime();
 	
 	public static void getTimeStarted( ) {
 		System.out.println("timeStarted is initially at" + timeStarted);
 	}
-	
-	public static float getTime() { return (float) ((System.nanoTime() - timeStarted) * 1E-9); }
+	// return in second
+	public static double getTime() { return (double) ((System.nanoTime() - timeStarted) * 1E-9); }
 	
 }
